reorganize pages and add styling for home site
