+++ src/Main.js
@@ -2,13 +2,13 @@ import React from 'react';
 import { Switch, Route } from 'react-router-dom';
 
 import Home from './pages/Home/Home';
-import TIForm from './pages/TIForm/TIForm';
+import TIApp from './pages/TI/TIApp';
 
 const Main = () => {
   return (
     <Switch> {/* The Switch decides which component to show based on the current URL.*/}
       <Route exact path='/' component={Home}></Route>
-      <Route exact path='/ti' component={TIForm}></Route>
+      <Route exact path='/ti' component={TIApp}></Route>
     </Switch>
   );
 }
