+++ src/pages/Home/Body.scss
@@ -52,6 +52,7 @@ color: var(--theme-dark-text);
 .grid-items {
   display: flex;
   flex-flow: row wrap;
+  box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2), 0 6px 20px 0 rgba(0, 0, 0, 0.19);
 }
 
 .grid-section-header {
