+++ src/pages/TI/TIApp.scss
@@ -1,0 +1,10 @@
+.ti-app {
+    height: 100vh;
+    background: #222222; 
+}
+.ti-app h1 {
+    margin: 0 0 30px;
+    padding: 0;
+    color: #fff;
+    text-align: center;
+}
