+++ src/pages/TI/TIForm.scss
@@ -1,10 +1,3 @@
-.ti-form h1 {
-    margin: 0 0 30px;
-    padding: 0;
-    color: rgba(0,0,0,.5);
-    text-align: center;
-}
-
 fieldset {
     border: none;
 }
